[feature/19-trustscore-api] rename : 신뢰점수 관련 패키지
