+++ src/main/java/com/example/demo/controller/trustscore/TrustScoreTypeController.java
@@ -1,4 +1,4 @@
-package com.example.demo.controller;
+package com.example.demo.controller.trustscore;
 
 import org.springframework.web.bind.annotation.RestController;
