+++ src/main/java/com/example/demo/repository/trustscore/TrustScoreHistoryRepository.java
@@ -1,6 +1,6 @@
-package com.example.demo.repository;
+package com.example.demo.repository.trustscore;
 
-import com.example.demo.model.TrustScoreHistory;
+import com.example.demo.model.trustscore.TrustScoreHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TrustScoreHistoryRepository extends JpaRepository<TrustScoreHistory, Long>, TrustScoreHistoryRepositoryCustom {
