+++ src/main/java/com/example/demo/repository/trustscore/TrustScoreRepository.java
@@ -1,6 +1,6 @@
-package com.example.demo.repository;
+package com.example.demo.repository.trustscore;
 
-import com.example.demo.model.TrustScore;
+import com.example.demo.model.trustscore.TrustScore;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TrustScoreRepository extends JpaRepository<TrustScore, Long>, TrustScoreRepositoryCustom {
