+++ src/main/java/com/example/demo/repository/trustscore/TrustScoreTypeRepository.java
@@ -1,6 +1,6 @@
-package com.example.demo.repository;
+package com.example.demo.repository.trustscore;
 
-import com.example.demo.model.TrustScoreType;
+import com.example.demo.model.trustscore.TrustScoreType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TrustScoreTypeRepository extends JpaRepository<TrustScoreType, Long>, TrustScoreTypeRepositoryCustom
