+++ src/main/java/com/example/demo/repository/trustscore/TrustScoreTypeRepositoryCustom.java
@@ -1,4 +1,4 @@
-package com.example.demo.repository;
+package com.example.demo.repository.trustscore;
 
 public interface TrustScoreTypeRepositoryCustom {
 }
