+++ src/main/java/com/example/demo/repository/trustscore/TrustScoreTypeRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.example.demo.repository;
+package com.example.demo.repository.trustscore;
 
 public class TrustScoreTypeRepositoryImpl implements TrustScoreTypeRepositoryCustom {
 }
