--- src/main/java/com/example/demo/service/TrustScoreTypeService.java
@@ -1,4 +1,0 @@
-package com.example.demo.service;
-
-public interface TrustScoreTypeService {
-}
