+++ src/main/java/com/example/demo/service/trustscore/TrustScoreHistoryService.java
@@ -1,4 +1,4 @@
-package com.example.demo.service;
+package com.example.demo.service.trustscore;
 
 public interface TrustScoreHistoryService {
 }
