+++ src/main/java/com/example/demo/service/trustscore/TrustScoreTypeService.java
@@ -1,0 +1,4 @@
+package com.example.demo.service.trustscore;
+
+public interface TrustScoreTypeService {
+}
