+++ src/main/java/com/example/demo/service/trustscore/impl/TrustScoreTypeServiceImpl.java
@@ -1,6 +1,7 @@
-package com.example.demo.service;
+package com.example.demo.service.trustscore.impl;
 
-import com.example.demo.repository.TrustScoreHistoryRepository;
+import com.example.demo.repository.trustscore.TrustScoreHistoryRepository;
+import com.example.demo.service.trustscore.TrustScoreTypeService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
